Refractor of routes
