+++ routes.js
@@ -18,6 +18,7 @@ const cityRoutes = require('./routes/cityRoutes.js');
 const tourRoutes = require('./routes/tourRoutes.js');
 const employeeRoutes = require('./routes/employeeRoutes.js');
 const bookingRoutes = require('./routes/bookingRoutes.js');
+const activitiesRoutes = require('./routes/activitiesRoutes.js');
 
 module.exports = function(app, express, db, log) {
 	if (log === undefined) {
@@ -91,146 +92,8 @@ module.exports = function(app, express, db, log) {
 	});
 
 	cityRoutes(app, db);
-
-	app.get('/api/activities', (req, res) => {
-		let cityName = req.query.city;
-		let price = req.query.price;
-		db.City.find({where: {name: cityName}}).then((city) => {
-			if (!city) {
-				res.status(404).send(JSON.stringify('no tours available in this city'));
-			} else {
-				db.Tour.findAll({where: {cityId: city.id}}).then((tourList) => {
-					if (tourList.length < 1) {
-						res.status(404).send(JSON.stringify('no tours available in this city'));
-					} else {
-						let toursArray = [];
-
-						for (var tour of tourList) {
-							if (tour.price <= price) {
-								var newTourInfo = {
-									title: tour.title,
-									description: tour.description,
-									city: cityName,
-									mainImage: tour.mainImage,
-									price: tour.price
-								}
-								toursArray.push(newTourInfo);
-							}
-						}
-						res.json(toursArray).end();
-					}
-				}).catch((error) => {res.status(500).send(JSON.stringify('error fetching tours'))});
-			}
-		}).catch((error) => {res.status(500).send(JSON.stringify('error fetching city information'))});
-	});
-
+	activitiesRoutes(app, db);
 	bookingRoutes(app, db);
-
-	// app.get('/api/bookings', (req, res) => {
-	// 	// console.log('bookings request...', req.query);
-	//   let tourId = req.query.tourId;
-	//   let date = req.query.date;
-	//   let userId = req.query.userId;
-	//   if (!tourId && !date && !!userId) {
-	//   	db.UserData.find({where: {userAuthId: userId}}).then((user) => {
-	//   		db.Booking.findAll({where: {touristId: user.id}}).then((bookingsList) => {
-	//   			let retArr = [];
-	//   			let asyncActions = [];
-
-	//   			for (var booking of bookingsList) {
-	//   				asyncActions.push(helpers.fillBookingData(booking, db).then((newBooking) => {
-	//   					retArr.push(newBooking);
-	//   				}));
-	//   			}
-
-	//   			Promise.all(asyncActions).then(() => {
-	//   				res.json(retArr).end();
-	//   			});
-	//   		})
-	//   	});
-	//   } else if (!tourId || !date || !userId) {
-	//     res.status(400).send(JSON.stringify('Invalid query string'));
-	//   } else {
-	//   	// console.log('getting tour..');
-	//     db.Tour.find({where: {id: tourId}}).then((tour) =>  {
-	//       if (!tour) {
-	//         res.status(404).send(JSON.stringify('Tour not found'));
-	//       } else {
-	//         db.City.find({where: {id: tour.dataValues.cityId}}).then((city) => {
-	//           if(!city) {
-	//             res.status(404).send(JSON.stringify('City not found'));
-	//           } else {
-	//             let booking = {tour: tour, city: city, date: date};
-	//             var driverOffering, guideOffering;
-	//             let findDriver = db.Offering.find({where: {cityId: city.dataValues.id, userType: 'Driver', date: date}}).then((driver) => {
-	//               if (driver) {
-	//                 driverOffering = driver;
-	//               }
-	//             });
-	//             let findGuide = db.Offering.find({where: {cityId: city.dataValues.id, userType: 'Tour Guide', date: date}}).then((guide) => {
-	//               if (guide) {
-	//                 guideOffering = guide;
-	//               }
-	//             });
-
-	//             Promise.all([findDriver, findGuide]).then(() => {
-	//               if (!!guideOffering && !!driverOffering) {
-
-	//               	let asyncActions = [];
-	//               	asyncActions.push(db.UserData.find({where: {id: driverOffering.userId}}).then((user) => {
-	//               		booking.driver = user;
-	//               	}));
-	//               	asyncActions.push(db.UserData.find({where: {id: guideOffering.userId}}).then((user) => {
-	//               		booking.guide = user;
-	//               	}));
-	//               	asyncActions.push(db.EmployeeData.find({where: {id: driverOffering.userId}}).then((user) => {
-	//               		booking.driverEmployeeData = user;
-	//               	}));
-	//               	asyncActions.push(db.EmployeeData.find({where: {id: guideOffering.userId}}).then((user) => {
-	//               		booking.guideEmployeeData = user;
-	//               	}));
-	//               	asyncActions.push(db.Offering.destroy({where: {id: driverOffering.id}}));
-	//               	asyncActions.push(db.Offering.destroy({where: {id: guideOffering.id}}));
-	//               	Promise.all(asyncActions).then(() => {
-	//                		db.UserData.find({where: {userAuthId: userId}}).then((user) => {
-	//                			if (!user) {
-	//                				res.status(400).send(JSON.stringify('user does not exist'));
-	//                			} else {
-	// 			              	let tourName = booking.tour.dataValues.title;
-	// 			              	let destinataries = [
-	// 												booking.driver.dataValues,
-	// 												booking.guide.dataValues,
-	// 												user.dataValues
-	// 			              	];
-	// 			              	mailer.sendMailToAll(destinataries, tourName, booking.date).then(function(response){
-	// 			              		// console.log('mail response', response);
-	// 			              	}, function(error) {
-	// 			              		// console.log(error)
-	// 			              	});
-	// 	               			db.Booking.create({
-	// 	               				driverId: booking.driver.id,
-	// 	               				touristId: user.id,
-	// 	               				tourGuideId: booking.guide.id,
-	// 	               				tourId: tourId,
-	// 	               				passengers: req.query.seats || null,
-	// 	               				date: date,
-	// 	               				cityId: booking.city.id
-	// 	               			});
-	// 	               			res.json(booking).end();
-	//                			}
-	//                		})
-	//               	});
-
-	//               } else {
-	//                 res.status(400).send(JSON.stringify('We were unable to book you with the given parameters'));
-	//               }
-	//             })
-	//           }
-	//         });
-	//       }
-	//     });
-	//   }
-	// });
 
 	app.get('/api/images/:imageName', (req, res) => {
 	  let imageName = req.params.imageName;
